Rename jarabe/web to jarabe/webservice
Given that web is being used for the HTML5/Javascript work, I am
renaming jarabe/web to jarabe/webservice.
