+++ src/jarabe/webservice/Makefile.am
@@ -1,4 +1,4 @@
-sugardir = $(pythondir)/jarabe/web
+sugardir = $(pythondir)/jarabe/webservice
 sugar_PYTHON =			\
 	__init__.py		\
 	account.py		\
